+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added publish checklist
 - Add CI
 - Add image to README.md
+- Don't show the complete button when there is no task to complete
 
 ## 2023-01-02 v0.1.0